clap-utils: Refactor compute_unit_price into compute_budget
